+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Changed
 
 - Removed `--disable` script flag - create data file with `Disabled` property set to `true` instead
+- Syncback `ignoreGlobs` now match directories with `/**` suffix (not only its contents)
 
 ## [2.0.19] - 2024-09-19